Refactoring of Kalman filter measurement updates.
It moves the performUpdate() and update*Noise() functions from LRKF to KF. This also
includes setUseAnalyticMeasurementModel() and getUseAnalyticMeasurementModel(). Subclasses
of KF now have to implement the getMomentFunc*() methods. All affected filters and
interfaces are updated according to all these changes. This allows for a much better
understanding of the code, code reuse is improved, and also the EKF (that did not inherit
from LRKF) can now use analytic moment computation for the update.